requirements.txt
aniso8601==9.0.1
astroid==2.5.6
attrs==21.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycparser==2.20
pylint==2.8.2
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
